blog/requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
django-crispy-forms==1.11.2
filelock==3.12.3
Pillow==10.0.0
platformdirs==3.10.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.4
bootstrap4